requirements.txt
# Testing & linting
pytest>=8.0
pytest-asyncio>=0.23
ruff>=0.4.0

# Dependencies
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.8
langchain-core==1.2.8
langchain-openai==1.1.7
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.8
MarkupSafe==3.0.3
MouseInfo==0.1.3
numpy==2.4.2
openai==2.16.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pillow==12.1.0
playwright==1.58.0
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.2.1
pytweening==1.2.0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rubicon-objc==0.5.3
sniffio==1.3.1
sounddevice==0.5.5
stagehand==3.6.0
starlette==0.52.1
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.41.0
xxhash==3.6.0
zstandard==0.25.0
